backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dj-rest-auth==7.0.1
django==4.2.21
django-allauth==65.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
pyjwt==2.10.1
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
